+++ lib/Utility/src/vec3.cpp
@@ -67,7 +67,7 @@ const float& vec3::operator[](const uint index) const {
 }
 
 bool vec3::operator==(const vec3 &v) const {
-    return std::abs(x - v.x) < VEC_EPS && std::abs(y - v.y) < VEC_EPS && std::abs(z - v.z) < VEC_EPS;
+    return std::abs(static_cast<int>(x - v.x)) < VEC_EPS && std::abs(static_cast<int>(y - v.y)) < VEC_EPS && std::abs(static_cast<int>(z - v.z)) < VEC_EPS;
 }
 
 std::ostream& operator<<(std::ostream& os, const vec3& v)
