+++ lib/Utility/src/vec4.cpp
@@ -85,8 +85,8 @@ const float& vec4::operator[](const uint index) const {
 }
 
 bool vec4::operator==(const vec4 &v) const {
-    return std::abs(x - v.x) < VEC_EPS && std::abs(y - v.y) < VEC_EPS &&
-           std::abs(z - v.z) < VEC_EPS && std::abs(w - v.w) < VEC_EPS;
+    return std::abs(static_cast<int>(x - v.x)) < VEC_EPS && std::abs(static_cast<int>(y - v.y)) < VEC_EPS &&
+           std::abs(static_cast<int>(z - v.z)) < VEC_EPS && std::abs(static_cast<int>(w - v.w)) < VEC_EPS;
 }
 
 bool vec4::operator!=(const vec4 &v) const {
